refactor: updated hook folder (#3224)
