+++ explorer/app/viewModelBuilders/azul/hca-dcp/common/viewModelBuilders.ts
@@ -41,9 +41,9 @@ import { METADATA_KEY } from "../../../../components/Index/common/entities";
 import { getPluralizedMetadataLabel } from "../../../../components/Index/common/indexTransformer";
 import { formatCountSize } from "../../../../components/Index/common/utils";
 import * as MDX from "../../../../content/hca-dcp";
-import { useExportEntityToTerraRequestParams } from "../../../../hooks/hca-dcp/useExportEntityToTerraRequestParams";
-import { useExportEntityToTerraRequestURL } from "../../../../hooks/hca-dcp/useExportEntityToTerraRequestURL";
-import { useExportEntityToTerraResponseURL } from "../../../../hooks/hca-dcp/useExportEntityToTerraResponseURL";
+import { useExportEntityToTerraRequestParams } from "../../../../hooks/azul/useExportEntityToTerraRequestParams";
+import { useExportEntityToTerraRequestURL } from "../../../../hooks/azul/useExportEntityToTerraRequestURL";
+import { useExportEntityToTerraResponseURL } from "../../../../hooks/azul/useExportEntityToTerraResponseURL";
 import { humanFileSize } from "../../../../utils/fileSize";
 import { mapAccessions } from "./accessionMapper/accessionMapper";
 import { Accession } from "./accessionMapper/entities";
